requirements.txt
PyQt5==5.15.11
PyQt5_sip==12.17.0
pyserial==3.5
